backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
numpy
openai
pydantic
python-multipart
httpx
groq
